fix: import and ES Module in evolucaoClinicaModel
